requirements.txt
Django==4.2.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
django-q==1.3.9
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
httpie==3.2.2
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1